backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic-settings>=2.6.0

# Gmail API
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.150.0

# Slack
slack-sdk>=3.33.0

# HTTP
httpx>=0.27.0
